+++ picoserve/src/time.rs
@@ -1,26 +1,50 @@
 //! [Timer] for creating timeouts during request parsing and request handling.
 
+#[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct TimeoutError;
+
+impl crate::io::Error for TimeoutError {
+    fn kind(&self) -> crate::io::ErrorKind {
+        crate::io::ErrorKind::TimedOut
+    }
+}
+
 /// A timer which can be used to abort futures if they take to long to resolve.
 pub trait Timer<Runtime> {
     /// The measure of time duration for this timer.
     type Duration: Clone;
-    /// The error returned if a future fails to resolve in time.
-    type TimeoutError;
+
+    async fn delay(&self, duration: Self::Duration);
 
     /// Drive the future, failing if it takes to long to resolve.
     async fn run_with_timeout<F: core::future::Future>(
         &self,
         duration: Self::Duration,
         future: F,
-    ) -> Result<F::Output, Self::TimeoutError>;
+    ) -> Result<F::Output, TimeoutError>;
 }
 
 pub(crate) trait TimerExt<Runtime>: Timer<Runtime> {
+    async fn timeout(&self, duration: Self::Duration) -> TimeoutError {
+        self.delay(duration).await;
+
+        TimeoutError
+    }
+
+    async fn maybe_timeout(&self, duration: Option<Self::Duration>) -> TimeoutError {
+        if let Some(duration) = duration {
+            self.timeout(duration).await
+        } else {
+            core::future::pending().await
+        }
+    }
+
     async fn run_with_maybe_timeout<F: core::future::Future>(
         &self,
         duration: Option<Self::Duration>,
         future: F,
-    ) -> Result<F::Output, Self::TimeoutError> {
+    ) -> Result<F::Output, TimeoutError> {
         if let Some(duration) = duration {
             self.run_with_timeout(duration, future).await
         } else {
@@ -38,14 +62,19 @@ pub struct TokioTimer;
 #[cfg(any(feature = "tokio", test))]
 impl Timer<super::TokioRuntime> for TokioTimer {
     type Duration = std::time::Duration;
-    type TimeoutError = tokio::time::error::Elapsed;
+
+    async fn delay(&self, duration: Self::Duration) {
+        tokio::time::sleep(duration).await
+    }
 
     async fn run_with_timeout<F: core::future::Future>(
         &self,
         duration: Self::Duration,
         future: F,
-    ) -> Result<F::Output, Self::TimeoutError> {
-        tokio::time::timeout(duration, future).await
+    ) -> Result<F::Output, TimeoutError> {
+        tokio::time::timeout(duration, future)
+            .await
+            .map_err(|_: tokio::time::error::Elapsed| TimeoutError)
     }
 }
 
@@ -56,14 +85,19 @@ pub struct EmbassyTimer;
 #[cfg(feature = "embassy")]
 impl Timer<super::EmbassyRuntime> for EmbassyTimer {
     type Duration = embassy_time::Duration;
-    type TimeoutError = embassy_time::TimeoutError;
+
+    async fn delay(&self, duration: Self::Duration) {
+        embassy_time::Timer::after(duration).await
+    }
 
     async fn run_with_timeout<F: core::future::Future>(
         &self,
         duration: Self::Duration,
         future: F,
-    ) -> Result<F::Output, Self::TimeoutError> {
-        embassy_time::with_timeout(duration, future).await
+    ) -> Result<F::Output, TimeoutError> {
+        embassy_time::with_timeout(duration, future)
+            .await
+            .map_err(|_: embassy_time::TimeoutError| TimeoutError)
     }
 }
 
@@ -87,7 +121,7 @@ impl<Runtime, W: crate::io::Write, T: Timer<Runtime>> crate::io::Write
         self.timer
             .run_with_maybe_timeout(self.timeout_duration.clone(), self.inner.write(buf))
             .await
-            .map_err(|_| super::Error::WriteTimeout)?
+            .map_err(super::Error::WriteTimeout)?
             .map_err(super::Error::Write)
     }
 
@@ -95,7 +129,7 @@ impl<Runtime, W: crate::io::Write, T: Timer<Runtime>> crate::io::Write
         self.timer
             .run_with_maybe_timeout(self.timeout_duration.clone(), self.inner.flush())
             .await
-            .map_err(|_| super::Error::WriteTimeout)?
+            .map_err(super::Error::WriteTimeout)?
             .map_err(super::Error::Write)
     }
 }